+++ change_log.md
@@ -1,5 +1,8 @@
 #Change log versions
-
+<!-- ## 1.2dev -->
+## 1.1
+- minor changes to display standards and physicsConstants
+- added gears and struct_meca
 ## 1.0.13
 - added pytest_util
 - dsp.stddisp bug fix inset